Move initialization to the Django hook, fix subtle bug with vectorize, remove the algo into their own algo/ folder
