+++ mangaki/mangaki/algo/__init__.py
@@ -1,0 +1,12 @@
+from .recommendation_algorithm import RecommendationAlgorithm, register_algorithm
+from .dataset import Dataset
+from .fit_algo import get_algo_backup, get_dataset_backup, fit_algo
+
+from .als import MangakiALS
+from .efa import MangakiEFA
+from .knn import MangakiKNN
+from .nmf import MangakiNMF
+from .pca import MangakiPCA
+from .svd import MangakiSVD
+from .wals import MangakiWALS
+from .zero import MangakiZero
