+++ mangaki/mangaki/algo/fit_algo.py
@@ -1,5 +1,7 @@
-from mangaki.utils.algos.recommendation_algorithm import RecommendationAlgorithm
-from mangaki.utils.algos.dataset import Dataset
+from .dataset import Dataset
+
+from mangaki.algo.recommendation_algorithm import RecommendationAlgorithm
+
 
 def fit_algo(algo_name, triplets, titles=None, categories=None, output_csv=False):
     algo = RecommendationAlgorithm.instantiate_algorithm(algo_name)
