+++ mangaki/mangaki/algo/nmf.py
@@ -1,11 +1,12 @@
-from django.conf import settings
-
-from mangaki.utils.algos.recommendation_algorithm import RecommendationAlgorithm
-from sklearn.decomposition import NMF
-from scipy.sparse import lil_matrix
-import numpy as np
 import csv
 import os
+
+import numpy as np
+from django.conf import settings
+from scipy.sparse import lil_matrix
+from sklearn.decomposition import NMF
+
+from mangaki.algo.recommendation_algorithm import RecommendationAlgorithm
 
 PIG_ID = 1124#1407 # QCTX=1434  JJ=1407  SebNL=1124
