+++ mangaki/mangaki/algo/pca.py
@@ -1,7 +1,7 @@
-from sklearn.decomposition import PCA
 import numpy as np
+from sklearn.decomposition import PCA
 
-from mangaki.utils.algos.recommendation_algorithm import RecommendationAlgorithm
+from mangaki.algo.recommendation_algorithm import RecommendationAlgorithm
 
 
 class MangakiPCA(RecommendationAlgorithm):
