+++ mangaki/mangaki/algo/wals.py
@@ -1,6 +1,8 @@
-from mangaki.utils.algos.recommendation_algorithm import RecommendationAlgorithm, register_algorithm
 from collections import defaultdict, Counter
+
 import numpy as np
+
+from mangaki.algo.recommendation_algorithm import RecommendationAlgorithm, register_algorithm
 
 
 def simple_train(model, inp, num_iterations):
