--- mangaki/mangaki/utils/__init__.py
@@ -1,4 +1,0 @@
-from .algos.recommendation_algorithm import RecommendationAlgorithm
-
-# Initialize the factory machinery at import-time.
-RecommendationAlgorithm.factory.initialize()
