+++ mangaki/mangaki/utils/dpp.py
@@ -1,7 +1,9 @@
-from mangaki.utils import dpplib
-from mangaki.utils.algos.fit_algo import get_algo_backup, get_dataset_backup
 from random import sample
+
 import numpy as np
+
+from mangaki.utils import dpplib
+from mangaki.algo import get_algo_backup, get_dataset_backup
 
 
 class MangakiUniform:
